refactor(node/fs): update fs import
